requirements.txt
-f /usr/share/pip-wheels
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django>=2.2.4
django-countries==5.3.3
django-embed-video==1.2.0
django-facebook==6.0.3
django-mysql==3.2.0
django-summernote==0.8.11.4
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.2
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.3
djangorestframework>=3.9.4
django-crispy-forms>=1.7.2
